.github/scripts/requirements.txt
requests>=2.31.0
supabase>=2.18.1
python-dotenv>=1.1.1
